+++ src/Components/Indicators/PageIndicator/PageIndicator.scss
@@ -1,4 +1,4 @@
-@import "../../colors";
+@import "../../../colors";
 
 .indicator {
     width: 70px;
